Test that groupFilter buttons exist and that they change query results
